+++ src/main/java/com/nbrichau/vanillaextension/stairs/RedstoneLampStairs.java
@@ -1,0 +1,59 @@
+package com.nbrichau.vanillaextension.stairs;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.StairsBlock;
+import net.minecraft.item.BlockItemUseContext;
+import net.minecraft.state.BooleanProperty;
+import net.minecraft.state.StateContainer;
+import net.minecraft.state.properties.BlockStateProperties;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import net.minecraft.world.server.ServerWorld;
+
+import java.util.Random;
+import java.util.function.Supplier;
+
+public class RedstoneLampStairs extends StairsBlock {
+
+	public static final BooleanProperty LIT = BlockStateProperties.LIT;
+
+	public RedstoneLampStairs(Supplier<BlockState> state, Properties properties) {
+		super(state, properties);
+		this.setDefaultState(this.getDefaultState().with(LIT, Boolean.FALSE));
+	}
+
+	@Override
+	public BlockState getStateForPlacement(BlockItemUseContext context) {
+		return super.getStateForPlacement(context).with(LIT, context.getWorld().isBlockPowered(context.getPos()));
+	}
+
+	@Override
+	public void neighborChanged(BlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos, boolean isMoving) {
+		if (!worldIn.isRemote) {
+			boolean flag = state.get(LIT);
+			if (flag != worldIn.isBlockPowered(pos)) {
+				if (flag) {
+					worldIn.getPendingBlockTicks().scheduleTick(pos, this, 4);
+				} else {
+					worldIn.setBlockState(pos, state.func_235896_a_(LIT), 2);
+				}
+			}
+
+		}
+	}
+
+	@Override
+	public void tick(BlockState state, ServerWorld worldIn, BlockPos pos, Random rand) {
+		if (state.get(LIT) && !worldIn.isBlockPowered(pos)) {
+			worldIn.setBlockState(pos, state.func_235896_a_(LIT), 2);
+		}
+
+	}
+
+	@Override
+	protected void fillStateContainer(StateContainer.Builder<Block, BlockState> builder) {
+		super.fillStateContainer(builder);
+		builder.add(LIT);
+	}
+}